backend/requirements.txt
asgiref==3.2.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==3.0.5
django-colorfield==0.6.3
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
drf-extra-fields==3.2.1
gunicorn==20.0.4
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.3
Pillow==9.0.0
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2020.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.3.1
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.8
psycopg2-binary==2.9.3
django-autoslug==1.9.8
